ALLJOYN-391 Add check for NULL pointer before dereferencing
If GetEnvironmentStrings is unsuccesful it will return a NULL pointer which will result in the var pointer also being NULL.  The code now check to see if the var pointer is NULL before dereferencing the pointer.  If the pointer is NULL Environ::Preload will log an error and return without trying to process the Environment Strings.