ALLJOYN-608 test case will not report failure on android
Moved the test case that tried to create a directory on root to the list of faulure cases since on Linux and Android if we try to create a directory or write a file to root (i.e. '/') we will get a 'Permission denied' error.  I changed the test to expect this error when running on any OS other than Windows.