Fix for ALLJOYN-1332
Fixing the corner-case of StringToDouble function in StringUtil.cc

a. There is no way to distinguish between the return value NaN and -NaN
   Hence, returning NaN in all cases.
b. Off-by-one error (accepting hex character A in input)
   The function uses decimal base (implicitly) and hence should NOT
   accept hexa-decimal character 'A' or '-a'.
c. Sanity check the exponent portion before using it
   The current state is to 'silently ignore' any incompatible input
   by taking the value of incompatible exponent portion as zero.
   To be consistent with above change 'b.', added a sanity check.

Enabled the DISABLED test as it now passes with the changes.