Fix for ALLJOYN-1399
The SocketTest.send_and_receive_with_dummy_fds ran to time limit
on Gingerbread emulator.

Root Cause:
NOT checking for the return value of Socket creation, led to an infinite
while loop. Gingerbread does NOT support IPv6 sockets.

Note: The while loop in itself isn't evil. It just runs till it finds
a free random port to bind the socket. Given the number of private ports
is about 16000, the chances of NOT finding 16 free ports are practically zero.